move libraries out of core
