+++ src/libraries/IActivities.ts
