+++ src/libraries/ICandles.ts
