+++ src/libraries/ICreds.ts
