+++ src/libraries/IEquitySymbols.ts
