+++ src/libraries/IQuotes.ts
