+++ src/libraries/IStockSymbol.ts
