+++ src/libraries/QtApi.ts
