+++ src/libraries/void0.ts
